# Conflicts:
#	api/client/event/event_stream.go
#	beacon-chain/blockchain/service.go
#	beacon-chain/blockchain/testing/mock.go
#	beacon-chain/core/feed/operation/events.go
#	beacon-chain/forkchoice/doubly-linked-tree/types.go
#	beacon-chain/forkchoice/interfaces.go
#	beacon-chain/forkchoice/ro_test.go
#	beacon-chain/node/node.go
#	beacon-chain/rpc/endpoints.go
#	beacon-chain/rpc/prysm/v1alpha1/validator/proposer.go
#	beacon-chain/rpc/prysm/v1alpha1/validator/server.go
#	beacon-chain/rpc/service.go
#	beacon-chain/sync/service.go
#	config/features/config.go
#	config/features/flags.go
#	config/params/config.go
#	config/params/mainnet_config.go
#	consensus-types/blocks/BUILD.bazel
#	proto/engine/v1/BUILD.bazel
#	proto/prysm/v1alpha1/BUILD.bazel
#	proto/prysm/v1alpha1/partial_data_columns.pb.go
#	testing/validator-mock/validator_mock.go
#	validator/client/health_monitor.go
#	validator/client/iface/validator.go
#	validator/client/testutil/mock_validator.go
#	validator/client/validator.go
#	validator/node/node.go
#	validator/rpc/server.go

+++ beacon-chain/state/stategen/setter.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/OffchainLabs/prysm/v7/beacon-chain/core/helpers"
 	"github.com/OffchainLabs/prysm/v7/beacon-chain/state"
+	"github.com/OffchainLabs/prysm/v7/config/features"
 	"github.com/OffchainLabs/prysm/v7/config/params"
 	"github.com/OffchainLabs/prysm/v7/encoding/bytesutil"
 	"github.com/OffchainLabs/prysm/v7/monitoring/tracing/trace"
@@ -47,7 +48,7 @@ func (s *State) ForceCheckpoint(ctx context.Context, blockRoot []byte) error {
 
 // This saves a post beacon state. On the epoch boundary,
 // it saves a full state. On an intermediate slot, it saves a back pointer to the
-// nearest epoch boundary state. If state-diff mode is enabled, it skips saving to hot state db.
+// nearest epoch boundary state.
 func (s *State) saveStateByRoot(ctx context.Context, blockRoot [32]byte, st state.BeaconState) error {
 	ctx, span := trace.StartSpan(ctx, "stateGen.saveStateByRoot")
 	defer span.End()
@@ -57,16 +58,32 @@ func (s *State) saveStateByRoot(ctx context.Context, blockRoot [32]byte, st stat
 
 	s.saveHotStateDB.lock.Lock()
 	if s.saveHotStateDB.enabled && st.Slot().Mod(duration) == 0 {
-		if err := s.beaconDB.SaveState(ctx, st, blockRoot); err != nil {
-			s.saveHotStateDB.lock.Unlock()
-			return err
-		}
-		s.saveHotStateDB.blockRootsOfSavedStates = append(s.saveHotStateDB.blockRootsOfSavedStates, blockRoot)
+		if features.Get().EnableStateDiff {
+			saver, ok := s.beaconDB.(hotStateSnapshotSaver)
+			if !ok {
+				s.saveHotStateDB.lock.Unlock()
+				return fmt.Errorf("hot state snapshot saver not supported")
+			}
+			if err := saver.SaveHotStateSnapshot(ctx, st, blockRoot); err != nil {
+				s.saveHotStateDB.lock.Unlock()
+				return err
+			}
 
-		log.WithFields(logrus.Fields{
-			"slot":                   st.Slot(),
-			"totalHotStateSavedInDB": len(s.saveHotStateDB.blockRootsOfSavedStates),
-		}).Info("Saving hot state to DB")
+			log.WithFields(logrus.Fields{
+				"slot": st.Slot(),
+			}).Info("Saving hot state to DB")
+		} else {
+			if err := s.beaconDB.SaveState(ctx, st, blockRoot); err != nil {
+				s.saveHotStateDB.lock.Unlock()
+				return err
+			}
+			s.saveHotStateDB.blockRootsOfSavedStates = append(s.saveHotStateDB.blockRootsOfSavedStates, blockRoot)
+
+			log.WithFields(logrus.Fields{
+				"slot":                   st.Slot(),
+				"totalHotStateSavedInDB": len(s.saveHotStateDB.blockRootsOfSavedStates),
+			}).Info("Saving hot state to DB")
+		}
 	}
 	s.saveHotStateDB.lock.Unlock()
 
@@ -156,17 +173,40 @@ func (s *State) DisableSaveHotStateToDB(ctx context.Context) error {
 		return nil
 	}
 
-	log.WithFields(logrus.Fields{
-		"enabled":          s.saveHotStateDB.enabled,
-		"deletedHotStates": len(s.saveHotStateDB.blockRootsOfSavedStates),
-	}).Warn("Exiting mode to save hot states in DB")
+	s.saveHotStateDB.enabled = false
 
 	// Delete previous saved states in DB as we are turning this mode off.
-	s.saveHotStateDB.enabled = false
-	if err := s.beaconDB.DeleteStates(ctx, s.saveHotStateDB.blockRootsOfSavedStates); err != nil {
-		return err
+	if features.Get().EnableStateDiff {
+		log.WithFields(logrus.Fields{
+			"enabled": s.saveHotStateDB.enabled,
+		}).Warn("Exiting mode to save hot states in DB")
+
+		clearer, ok := s.beaconDB.(hotStateSnapshotClearer)
+		if !ok {
+			return fmt.Errorf("hot state snapshot clearer not supported")
+		}
+		if err := clearer.ClearHotStateSnapshots(ctx); err != nil {
+			return err
+		}
+	} else {
+		log.WithFields(logrus.Fields{
+			"enabled":          s.saveHotStateDB.enabled,
+			"deletedHotStates": len(s.saveHotStateDB.blockRootsOfSavedStates),
+		}).Warn("Exiting mode to save hot states in DB")
+
+		if err := s.beaconDB.DeleteStates(ctx, s.saveHotStateDB.blockRootsOfSavedStates); err != nil {
+			return err
+		}
 	}
 	s.saveHotStateDB.blockRootsOfSavedStates = nil
 
 	return nil
+}
+
+type hotStateSnapshotSaver interface {
+	SaveHotStateSnapshot(ctx context.Context, st state.ReadOnlyBeaconState, root [32]byte) error
+}
+
+type hotStateSnapshotClearer interface {
+	ClearHotStateSnapshots(ctx context.Context) error
 }
